Mate and slave summoning: handle, skill list and wire layout
The mate handle is what ties the world object to the mate record. A slave gets
that pairing from SCMySlave, which carries the object id and the handle
together; a mate has no such message, so the only place the two ever meet is
the mate branch of the unit state. That field was carrying the owner object id
truncated to 16 bits, so the record filed under the handle was unreachable: the
mount stood in the world with nothing behind it, an empty skill bar and nothing
to ride.

The ten skill slots of SCMateSpawned take the npc_mount_skills row id, not the
skill it resolves to. Resolving it first looked right because the return path
matches on the real skill id, but the client draws nothing from it.

Mate spawning now logs the template, the category and the ten values as they go
out, which is how both of the above were settled.

Also in this change, from the same body of work: the master unit id inside the
BaseUnitType branch is two bytes rather than three, a bonded slave was missing
the leading point byte of its second point context, and mate equipment slots
19-25 carry only the item type.

+++ AAEmu.Game/Core/Packets/G2C/SCMySlavePacket.cs
@@ -10,13 +10,13 @@ public class SCMySlavePacket : GamePacket
     private readonly ushort _tl; // TODO slaveId
     private readonly string _slaveName;
     private readonly uint _templateId;
-    private readonly int _hp;
-    private readonly int _maxHp;
+    private readonly long _hp;
+    private readonly long _maxHp;
     private readonly float _x;
     private readonly float _y;
     private readonly float _z;
 
-    public SCMySlavePacket(uint unitId, ushort tl, string slaveName, uint templateId, int hp, int maxHp, float x, float y, float z)
+    public SCMySlavePacket(uint unitId, ushort tl, string slaveName, uint templateId, long hp, long maxHp, float x, float y, float z)
         : base(SCOffsets.SCMySlavePacket, 5)
     {
         _unitId = unitId;
@@ -36,8 +36,8 @@ public class SCMySlavePacket : GamePacket
         stream.Write(_tl);           // slaveId
         stream.Write(_slaveName);    // slaveName
         stream.Write(_templateId);   // type
-        stream.Write(_hp);           // hp
-        stream.Write(_maxHp);        // maxHp
+        stream.Write(_hp);           // hp    : i64
+        stream.Write(_maxHp);        // maxHp : i64
         stream.Write(Helpers.ConvertLongX(_x)); // WorldPosXYZ_0940
         stream.Write(Helpers.ConvertLongY(_y));
         stream.Write(_z);
